+++ src/types/node-transformers/TNodeTransformersRelationEdge.d.ts
@@ -1,0 +1,3 @@
+import { NodeTransformer } from '../../enums/node-transformers/NodeTransformer';
+
+export type TNodeTransformersRelationEdge = [NodeTransformer, NodeTransformer | null];