+++ src/types/node-transformers/TNormalizedNodeTransformers.d.ts
@@ -1,0 +1,5 @@
+import { TObject } from '../TObject';
+
+import { INodeTransformer } from '../../interfaces/node-transformers/INodeTransformer';
+
+export type TNormalizedNodeTransformers = TObject<INodeTransformer>;