+++ test/dev/dev.ts
@@ -6,18 +6,16 @@ import { NO_ADDITIONAL_NODES_PRESET } from '../../src/options/presets/NoCustomNo
 
     let obfuscatedCode: string = JavaScriptObfuscator.obfuscate(
         `
-        function hi() {
-          console.log("Hello World!");
-        }
-        hi();
+        var s = {
+            'abcdefg': 'abcdefg'
+        };
         `,
         {
             ...NO_ADDITIONAL_NODES_PRESET,
             compact: false,
             splitStrings: true,
             splitStringsChunkLength: 4,
-            controlFlowFlattening: true,
-            controlFlowFlatteningThreshold: 1
+            transformObjectKeys: true
         }
     ).getObfuscatedCode();